updated add sales api
