+++ supermarket-action-service/src/main/java/com/eidiko/supermarket_action_service/services/SalesService.java
@@ -1,6 +1,6 @@
 package com.eidiko.supermarket_action_service.services;
 import com.eidiko.supermarket_action_service.dao.SalesRepo;
-import com.eidiko.supermarket_action_service.dto.SalesDto;
+import com.eidiko.supermarket_action_service.dto.SaleRequest;
 import com.eidiko.supermarket_action_service.model.Sales;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -16,9 +16,9 @@ public class SalesService {
        this.salesRepo=salesRepo;
    }
 
-   public Sales addSales(SalesDto salesDto)
+   public Sales addSales(SaleRequest saleRequest)
    {
-      return salesRepo.addSales(salesDto);
+      return salesRepo.addSales(saleRequest);
    }
 
 }
